+++ frontend/src/components/NFTItem/index.tsx
@@ -3,49 +3,16 @@ import { calculateERC721MechAddress } from "mech"
 
 import classes from "./NFTItem.module.css"
 import Button from "../Button"
-import { useEffect, useState } from "react"
+import { useState } from "react"
 import { shortenAddress } from "../../utils/shortenAddress"
+import useTokenUrl from "../../hooks/useTokenUrl"
 
 interface Props {
   nft: Nft
 }
 
-const useTokenUrl = (tokenUrl: string | undefined) => {
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<any>(false)
-  const [data, setData] = useState<any>(null)
-
-  useEffect(() => {
-    if (!tokenUrl) {
-      setIsLoading(false)
-      return
-    }
-
-    const fetchData = async () => {
-      if (tokenUrl.slice(0, 4) === "http") {
-        try {
-          const response = await fetch(tokenUrl)
-          const data = await response.json()
-          setData(data)
-          setIsLoading(false)
-        } catch (e) {
-          setError(e)
-          setIsLoading(false)
-        }
-      } else if (tokenUrl.slice(0, 29) === "data:application/json;base64,") {
-        const json = Buffer.from(tokenUrl.slice(29), "base64").toString()
-        setData(JSON.parse(json))
-        setIsLoading(false)
-      }
-    }
-
-    fetchData()
-  }, [tokenUrl])
-
-  return { isLoading, data, error }
-}
-
 const NFTItem: React.FC<Props> = ({ nft }) => {
+  const [imageError, setImageError] = useState(false)
   const needTokenUrl = !nft.imageUrl
   const { isLoading, data, error } = useTokenUrl(
     needTokenUrl ? nft.tokenUrl : undefined
@@ -61,12 +28,15 @@ const NFTItem: React.FC<Props> = ({ nft }) => {
         )}
       </div>
       <div className={classes.main}>
-        {error && <div className={classes.noImage}></div>}
-        {!isLoading && !error && (
+        {(error || imageError || isLoading) && (
+          <div className={classes.noImage}></div>
+        )}
+        {!isLoading && !error && !imageError && (
           <img
             src={data ? data.image : nft.imageUrl}
             alt={nft.name}
             className={classes.image}
+            onError={() => setImageError(true)}
           />
         )}
         <div className={classes.info}>
